完善了文档

+++ app/src/main/java/com/example/himalaya/interfaces/IPlayerCallback.kt
@@ -21,9 +21,11 @@ interface IPlayerCallback {
 
     fun onPlayModeChange(playMode: XmPlayListControl.PlayMode)
 
-    fun onProgressChange(currentIndex:Long,total:Long)
+    fun onProgressChange(currentIndex:Int,total:Int)
 
     fun onAdLoading()
 
     fun onAdFinished()
+
+
 }
